Converts dataframe from long/tidy to wide format with NAs removed

This conversion is helpful mostly for repeated measures design.

long_to_wide_converter(
  data,
  x,
  y,
  subject.id = NULL,
  paired = TRUE,
  spread = TRUE,
  ...
)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

subject.id
In case of repeated measures design (paired = TRUE, i.e.), this argument specifies the subject or repeated measures id. Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate.

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

spread
Logical that decides whether the dataframe needs to be converted from long/tidy to wide (default: TRUE), or the data needs to be returned as it is but with the NAs removed.

...
Currently ignored.

Value

A dataframe in the wide (or Cartesian) format.

Examples

# for reproducibility
library(ipmisc)
set.seed(123)

# repeated measures design
long_to_wide_converter(
  data = bugs_long,
  x = condition,
  y = desire,
  subject.id = subject,
  paired = TRUE
)

#> # A tibble: 88 x 5
#>    rowid  HDHF  HDLF  LDHF  LDLF
#>    <int> <dbl> <dbl> <dbl> <dbl>
#>  1     1  10     9     6     6  
#>  2     3  10    10    10     5  
#>  3     4   9     6     9     6  
#>  4     5   8.5   5.5   6.5   3  
#>  5     6   3     7.5   0.5   2  
#>  6     7  10    10    10    10  
#>  7     8  10     9    10    10  
#>  8     9  10     6     9.5   9.5
#>  9    11   0     0     2.5   0  
#> 10    12  10     8.5   7.5   9.5
#> # ... with 78 more rows

# independent measures design
long_to_wide_converter(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  paired = FALSE,
  spread = FALSE
)

#> # A tibble: 51 x 3
#>    rowid vore  brainwt
#>    <int> <fct>   <dbl>
#>  1     3 carni  0.07  
#>  2     4 carni  0.0108
#>  3     5 carni  0.0256
#>  4     7 carni  0.325 
#>  5     9 carni  0.0125
#>  6    12 carni  0.157 
#>  7    17 carni  0.0175
#>  8    18 carni  0.0445
#>  9    19 carni  0.0504
#> 10     2 herbi  0.423 
#> # ... with 41 more rows